requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
djangorestframework==3.12.4
folium==0.12.1
idna==2.10
install==1.3.4
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.20.1
pandas==1.2.3
protobuf==3.15.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.3
